+++ src/common/gitProvider/azureDevops.ts
@@ -11,7 +11,7 @@ export class AzureDevopsProvider extends GitProviderRoot {
   constructor() {
     super();
     // Azure server url must be provided in AZURE_SERVER_URL. ex: https:/dev.azure.com/mycompany
-    this.serverUrl = process.env.SYSTEM_COLLECTION_URI;
+    this.serverUrl = process.env.SYSTEM_COLLECTIONURI;
     // a Personal Access Token must be defined
     this.token = process.env.CI_SFDX_HARDIS_AZURE_TOKEN || process.env.SYSTEM_ACCESSTOKEN;
     const authHandler = azdev.getHandlerFromToken(this.token);
@@ -27,6 +27,7 @@ export class AzureDevopsProvider extends GitProviderRoot {
     // Get CI variables
     const repositoryId = process.env.BUILD_REPOSITORY_ID || null;
     const buildId = process.env.BUILD_BUILD_ID || null;
+    const jobId = process.env.SYSTEM_JOB_ID || null;
     const pullRequestIdStr = process.env.SYSTEM_PULLREQUEST_PULLREQUESTID || null;
     if (repositoryId == null || pullRequestIdStr == null) {
       uxLog(this, c.grey("[Azure integration] No project and pull request, so no note thread..."));
@@ -36,7 +37,7 @@ export class AzureDevopsProvider extends GitProviderRoot {
     const azureJobName = process.env.SYSTEM_JOB_DISPLAY_NAME;
     const SYSTEM_COLLECTIONURI = process.env.SYSTEM_COLLECTIONURI.replace(/ /g, "%20");
     const SYSTEM_TEAMPROJECT = process.env.SYSTEM_TEAMPROJECT.replace(/ /g, "%20");
-    const azureBuildUri = `${SYSTEM_COLLECTIONURI}${SYSTEM_TEAMPROJECT}/_build/results?buildId=${buildId}`;
+    const azureBuildUri = `${SYSTEM_COLLECTIONURI}${SYSTEM_TEAMPROJECT}/_build/results?buildId=${buildId}&view=logs&j=${jobId}`;
     // Build thread message
     const messageKey = prMessage.messageKey + "-" + azureJobName + "-" + pullRequestId;
     let messageBody = `**${prMessage.title || ""}**
@@ -44,6 +45,7 @@ export class AzureDevopsProvider extends GitProviderRoot {
 ${prMessage.message}
 
 <br/>
+
 _Provided by [sfdx-hardis](https://sfdx-hardis.cloudity.com) from job [${azureJobName}](${azureBuildUri})_
 <!-- sfdx-hardis message-key ${messageKey} -->
 `;
@@ -60,43 +62,45 @@ _Provided by [sfdx-hardis](https://sfdx-hardis.cloudity.com) from job [${azureJo
     let existingThreadComment: GitPullRequestCommentThread = null;
     let existingThreadCommentId: number = null;
     for (const existingThread of existingThreads) {
-      if (existingThread?.comments[0]?.content.includes(`<!-- sfdx-hardis message-key ${messageKey} -->`)) {
-        existingThreadComment = existingThread;
-        existingThreadCommentId = existingThread.comments[0].id;
-        existingThreadId = existingThread.id;
+      for (const comment of existingThread?.comments || []) {
+        if ((comment?.content || "").includes(`<!-- sfdx-hardis message-key ${messageKey} -->`)) {
+          existingThreadComment = existingThread;
+          existingThreadCommentId = existingThread.comments[0].id;
+          existingThreadId = existingThread.id;
+          break;
+        }
+      }
+      if (existingThreadId) {
+        break;
       }
     }
 
     // Create or update MR note
     if (existingThreadId) {
       // Delete previous comment
-      uxLog(this, c.grey("[Azure integration] Deleting previous comment..."));
-      const deletedCommentResult = await azureGitApi.deleteComment(repositoryId, pullRequestId, existingThreadId, existingThreadCommentId);
+      uxLog(this, c.grey("[Azure integration] Deleting previous comment and closing previous thread..."));
+      await azureGitApi.deleteComment(repositoryId, pullRequestId, existingThreadId, existingThreadCommentId);
+      existingThreadComment = await azureGitApi.getPullRequestThread(repositoryId, pullRequestId, existingThreadId);
       // Update existing thread
-      uxLog(this, c.grey("[Azure integration] Updating Pull Request Thread on Azure..."));
-      existingThreadComment.comments[0] = { content: messageBody };
-      existingThreadComment.status = this.pullRequestStatusToAzureThreadStatus(prMessage);
-      const azureEditThreadResult = await azureGitApi.updateThread(existingThreadComment, repositoryId, pullRequestId, existingThreadId);
-      const prResult: PullRequestMessageResult = {
-        posted: azureEditThreadResult.id > 0,
-        providerResult: azureEditThreadResult,
-        additionalProviderResult: { deletedCommentResult: deletedCommentResult },
+      existingThreadComment = {
+        id: existingThreadComment.id,
+        status: CommentThreadStatus.Closed,
       };
-      return prResult;
-    } else {
-      // Create new thread
-      uxLog(this, c.grey("[Azure integration] Adding Pull Request Thread on Azure..."));
-      const newThreadComment: GitPullRequestCommentThread = {
-        comments: [{ content: messageBody }],
-        status: this.pullRequestStatusToAzureThreadStatus(prMessage),
-      };
-      const azureEditThreadResult = await azureGitApi.createThread(newThreadComment, repositoryId, pullRequestId);
-      const prResult: PullRequestMessageResult = {
-        posted: azureEditThreadResult.id > 0,
-        providerResult: azureEditThreadResult,
-      };
-      return prResult;
+      await azureGitApi.updateThread(existingThreadComment, repositoryId, pullRequestId, existingThreadId);
     }
+
+    // Create new thread
+    uxLog(this, c.grey("[Azure integration] Adding Pull Request Thread on Azure..."));
+    const newThreadComment: GitPullRequestCommentThread = {
+      comments: [{ content: messageBody }],
+      status: this.pullRequestStatusToAzureThreadStatus(prMessage),
+    };
+    const azureEditThreadResult = await azureGitApi.createThread(newThreadComment, repositoryId, pullRequestId);
+    const prResult: PullRequestMessageResult = {
+      posted: azureEditThreadResult.id > 0,
+      providerResult: azureEditThreadResult,
+    };
+    return prResult;
   }
 
   // Convert sfdx-hardis PR status to Azure Thread status value